# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Migration Factory API",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Project",
    "operationId": "create_project_api_v1_projects_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Projects",
    "operationId": "list_projects_api_v1_projects_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ProjectOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/scans": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Scan",
    "operationId": "create_scan_api_v1_projects__project_id__scans_post",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScanCreate"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScanAccepted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Scans",
    "operationId": "list_scans_api_v1_projects__project_id__scans_get",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ScanOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "Get Scan",
    "operationId": "get_scan_api_v1_scans__scan_id__get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}/modules": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "Modules",
    "operationId": "modules_api_v1_scans__scan_id__modules_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ModuleOut"
         },
         "title": "Response Modules Api V1 Scans  Scan Id  Modules Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}/routes": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "Routes",
    "operationId": "routes_api_v1_scans__scan_id__routes_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     },
     {
      "name": "dsl",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(java|xml)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Dsl"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_RouteOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}/findings": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "Findings",
    "operationId": "findings_api_v1_scans__scan_id__findings_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     },
     {
      "name": "severity",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(high|medium|low|info)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Severity"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(open|proposed|resolved|discarded)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "rule",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^[A-Z0-9_]{2,60}$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Rule"
      }
     },
     {
      "name": "classification",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(AUTO|AUTO_TEST|REVIEW|MANUAL)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Classification"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 500
        },
        {
         "type": "null"
        }
       ],
       "title": "Module"
      }
     },
     {
      "name": "file",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 500
        },
        {
         "type": "null"
        }
       ],
       "title": "File"
      }
     },
     {
      "name": "blocking",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Blocking"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_FindingOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/findings/{finding_id}": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "Get Finding",
    "operationId": "get_finding_api_v1_findings__finding_id__get",
    "parameters": [
     {
      "name": "finding_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Finding Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FindingDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/findings/{finding_id}/review": {
   "patch": {
    "tags": [
     "scans"
    ],
    "summary": "Review Finding",
    "operationId": "review_finding_api_v1_findings__finding_id__review_patch",
    "parameters": [
     {
      "name": "finding_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Finding Id"
      }
     },
     {
      "name": "if-match",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string",
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FindingReviewIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FindingOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}/plans": {
   "post": {
    "tags": [
     "scans"
    ],
    "summary": "New Plan",
    "operationId": "new_plan_api_v1_scans__scan_id__plans_post",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PlanCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "List Plans",
    "operationId": "list_plans_api_v1_scans__scan_id__plans_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlanOut"
         },
         "title": "Response List Plans Api V1 Scans  Scan Id  Plans Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/permissions": {
   "get": {
    "tags": [
     "scans"
    ],
    "summary": "My Permissions",
    "operationId": "my_permissions_api_v1_projects__project_id__permissions_get",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Permissions"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ModuleOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "path": {
      "type": "string",
      "title": "Path"
     },
     "groupId": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Groupid"
     },
     "artifactId": {
      "type": "string",
      "title": "Artifactid"
     },
     "version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Version"
     },
     "packaging": {
      "type": "string",
      "title": "Packaging"
     },
     "javaObserved": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Javaobserved"
     },
     "javaResolved": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Javaresolved"
     },
     "javaSource": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Javasource"
     },
     "camelVersions": {
      "items": {},
      "type": "array",
      "title": "Camelversions"
     },
     "resolution": {
      "type": "string",
      "title": "Resolution"
     },
     "unresolved": {
      "items": {},
      "type": "array",
      "title": "Unresolved"
     },
     "dependencies": {
      "items": {},
      "type": "array",
      "title": "Dependencies"
     },
     "profiles": {
      "items": {},
      "type": "array",
      "title": "Profiles"
     }
    },
    "type": "object",
    "required": [
     "id",
     "path",
     "groupId",
     "artifactId",
     "version",
     "packaging",
     "javaObserved",
     "javaResolved",
     "javaSource",
     "camelVersions",
     "resolution",
     "unresolved",
     "dependencies",
     "profiles"
    ],
    "title": "ModuleOut"
   },
...
   "Page_ProjectOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ProjectOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "nextCursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nextcursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[ProjectOut]"
   },
...
   "Page_ScanOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ScanOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "nextCursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nextcursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[ScanOut]"
   },
...
   "ProjectCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "pattern": "^[\\w .\\-()]+$",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "maxLength": 2000,
      "title": "Description",
      "default": ""
     },
     "sourceType": {
      "type": "string",
      "enum": [
       "local",
       "zip",
       "git"
      ],
      "title": "Sourcetype"
     },
     "targetRuntime": {
      "type": "string",
      "enum": [
       "spring",
       "quarkus"
      ],
      "title": "Targetruntime"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "defaultBranch": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[\\w./-]{1,200}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Defaultbranch"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "uploadId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Uploadid"
     }
    },
    "type": "object",
    "required": [
     "name",
     "sourceType",
     "targetRuntime"
    ],
    "title": "ProjectCreate"
   },
   "ProjectOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "sourceType": {
      "type": "string",
      "title": "Sourcetype"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "repositoryProvider": {
      "type": "string",
      "title": "Repositoryprovider"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "upload": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Upload"
     },
     "defaultBranch": {
      "type": "string",
      "title": "Defaultbranch"
     },
     "targetRuntime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Targetruntime"
     },
     "requiredGates": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Requiredgates"
     },
     "independentReview": {
      "type": "boolean",
      "title": "Independentreview"
     },
     "prConfig": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prconfig"
     },
     "myRole": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Myrole"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "lastScan": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lastscan"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "description",
     "sourceType",
     "repositoryUrl",
     "repositoryProvider",
     "credentialRef",
     "defaultBranch",
     "targetRuntime",
     "requiredGates",
     "independentReview",
     "prConfig",
     "myRole",
     "version",
     "createdAt"
    ],
    "title": "ProjectOut"
   },
...
   "ScanAccepted": {
    "properties": {
     "scanId": {
      "type": "string",
      "format": "uuid",
      "title": "Scanid"
     },
     "jobId": {
      "type": "string",
      "format": "uuid",
      "title": "Jobid"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "scanId",
     "jobId",
     "status"
    ],
    "title": "ScanAccepted"
   },
   "ScanCreate": {
    "properties": {
     "ref": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[\\w./-]{1,200}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref"
     },
     "catalogId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Catalogid"
     },
     "target": {
      "type": "string",
      "enum": [
       "spring",
       "quarkus"
      ],
      "title": "Target",
      "default": "spring"
     },
     "projectRoot": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500,
        "pattern": "^[\\w.\\-]+(/[\\w.\\-]+)*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Projectroot"
     }
    },
    "type": "object",
    "title": "ScanCreate"
   },
...
  }
 }
}